adds $mOriginal array as a way to track the filename, mimetype, and filesize of
the original upload object.  That way $this->mFilename $this->mMimeType and
$this->mFilesize can be always used for the file being processed.
$mFileSize was changed to $mFilesize
mArticleTitle is created by a combination of a source file, and it's
mArticleSavePath when trying to re-upload a single file as part of a set.
mIsVIP and mIsUVM for identifying collection types
mImages for holding an array of image files to process.
$tidyOpts is now a static array for handling internally
to the class without a configuration file.
loadRequest() takes on an additional form field for the CollectionName
Adds a better working version of ::getMimeType() and uncovers bug in the
original js code where mimetype is treated like an extension.
doUpload now only sets mimetype, filename and filesize (plus puts those in
mOriginal) processFile() is concerned with gathering its contents.
reverted to the "old" way of declaring globals because the whole getConfig()->get()
was not working.
set a switch in doUpload to either unwrap, or processFile() based on the mimetype
add processImages() to the end of doUpload()
Adds unwrapZipFile()
Stubs showResults()
doImport() is now almost superfluous: doUpload() and showResults() or showForm()
processFile() is now all the things that used to happen in doUpload()
Adds processImages()
makeTitle() now operates on mArticleTitle instead of mFilename
tidyUp() now adds mimetype checking
tidyUp() now uses a HERESTRING to process this->mContent rather than a
file on disk - in the fallback case where we have to shell out to
a command line tidy.  The PHP tidy will always handle the content variable
Adds qpAddParentToLink() to turn indefinite relative links into properly referenced
links of a collection
::qpItalics() to turn HTML markup to mediawiki italics markup
Adds qpCleanLinks() to remove non-essential attributes; plus remove empty links
Adds qpRemoveMouseOvers()
Adds qpRewriteImages()to turn HTML image tags into MediaWiki markup
Fixes $stage and 'from' option to panDoc2Wiki()