UI enhancement - adds spinner; refactored link handling
To provide an indication that the application is working, there is now a
'spinner' that displays once you click the submit button.
Adds .mw-ext-Html2Wiki-loading class in ext.Html2Wiki.css

ext.Html2Wiki.js adds a simple jQuery click handler

$wgH2WEliminateDuplicateImages defaults to false; and we no longer
try to get fancy with eliminating 'images/' in collection paths.

removeExtensionFromPath() gets faster, and handles named anchors.  And
previewing the day when we're going to need to import non-HTML content
with Html2Wiki, we add the option to not affect paths that are non-html

qpNormalizeLinks() is new
qpMakeLinksAbsolute() and qpMakeLinksStatic() are deprecated

adds id to submit button for jQuery handler

adds early return checks for qp* functions based on length

qpAlterImageLinks() is more generic now due to QueryParse problem
searching for non-existant parent elements.
qpRewriteImages() is deleted
eliminateCruft() is deleted