Add feature: Dry-run option
The dry-run option allows you to preview the conversion without saving.
This is a prelude to being able to also manage the conversion with config
options or "recipes".  When the "dry-run" option is checked in the form,
the conversion results will be displayed, but not saved to the wiki.

New strings in the i18n.json files for both the form, and display results
Also tweaked the form text to describe support for multiple archive formats.

New property $mIsDryRun populated by the 'dry-run' form element (checkbox).

If $mIsDryRun is true we call Html2Wiki::showContent() instead of
Html2Wiki::saveArticle()